+++ src/view/UI/components/AddAgencyChip/useChipAnalytics.tsx
@@ -1,0 +1,15 @@
+import { useAnalytics } from "../../hooks";
+import { ECategories } from "../../../../libs";
+
+export const useChipAnalytics = () => {
+  const analytics = useAnalytics();
+
+  const event = () => {
+    analytics.click({
+      category: ECategories.ORDERS,
+      label: "add agency click",
+    });
+  };
+
+  return { event };
+};
